Sample environment with old ServiceNetMap value
The original value for the ServiceNetMap parameter had the Keystone
Admin API service on the Internal API network. Later, it was moved to
the ctlplane network by default.

Users updating from clouds already deployed may not want to have the
service moved, and we've occassionly seen it cause issues with services
not getting restarted properly.

This sample environment file documents the old value so that users can
just optionally include it via -e to keep the services the same as they
were when they originally deployed.

Change-Id: I0b68542337a2f40e26df15fe7ac2da5aafe651d5