Add SoftwareConfigTransport for switching transports
This change adds a SoftwareConfigTransport parameter to role templates
so that the transport can be changed via a parameter_defaults entry.

Changing the transport to POLL_TEMP_URL on an existing overcloud has
worked for local testing when heat change
I9c475f0c489c67db5895924050186228403f2773 is running on the undercloud.

Change-Id: I5c2a2d2170714093c5757282cba12ac65f8738a4