Add option to add metadata for the overcloud nodes
Some Nova hooks might require custom properties/metadata set for the
servers deployed in the overcloud, and this would enable us to inject
such information.

For FreeIPA (IdM) integration, there is effectively a Nova hook that
requires such data.

Currently this inserts metadata for all servers, but a subsequent CR
will introduce per-role metadata. However, that was not added to this
because it will require the usage of map_merge. which will block those
changes to be backported. However, this one is not a problem in that
sense.

Change-Id: I98b15406525eda8dff704360d443590260430ff0
(cherry picked from commit a7925bc9ed253baae87185016b65dc4cb774d966)
Conflicts:
	overcloud.yaml
	puppet/ceph-storage.yaml
	puppet/cinder-storage.yaml
	puppet/compute.yaml
	puppet/controller.yaml
	puppet/swift-storage.yaml