Add update yaml backward compatibe with PublicVirtualIP on ctlplane
In previous releases, when not using network isolation, we used to create
two different VIPs for the ControlVirtualIP and the PublicVirtualIP both on
the ctlplane network. Later we moved into a configuration with a single
VIP instead so we need a compatibility yaml for those updating from old
versions which preserves both the IPs; one of the two is deleted
otherwise.

Also updates README.md with a short description of the use case.

Change-Id: Iae08b938a255bf563d3df2fdc0748944a9868f8e
(cherry picked from commit 3770deda062b0bb3291fdfc57d7b5e000392f9ec)