Configure keystone public_endpoint
We need this set for SSL or keystone returns a non-https address.
It shouldn't hurt anything to set this in the non-SSL case since
the value will still be correct and the behavior will be the same
as if it were unset.

Change-Id: Iea3ea1d25dfc462fa844d3c12e6070f2c9b42036
(cherry picked from commit ae911c2c9c2f48f4a3278c47fe05fbf552bdaeca)