Minor fixes to allow local docker registry usage
Changed the heat-docker-agents namespace to use the namespacing
specified in the environment file, which reduces modifications required
on the user when using a local registry.

Changed the start agents script to handle using a local registry both
with a namespace and without.

Change-Id: I16cc96b7ecddeeda07de45f50ffc6a880dabbba6
(cherry picked from commit 255f4fd69e6bdf8247bc2861fa830859fffbef33)