Allow to enable IPv6 on Corosync
To enable IPv6 in Corosync, you'll need to set the --ipv6 parameter when
building the cluster.

To do so, we need to specify this parameter in puppet configuration when
running ::pacemaker::corosync class.

So we have now a new parameter: CorosyncIPv6 which is a boolean, False
by default. If set to True, it will pass right options to enable IPv6
when building the Pacemaker cluster.

Change-Id: I485f1399964ddd7a201ed19ec6580b8a136b3acd