Adds v6 capability to the deploy validation test (pings)
This changes the ping_controller_ips function in the all-nodes.sh
bash validation script which is run during deployment to check
network connectivity (to fail early).

The main differences are using the v6 routes when it is a v6 address
and using python -c to check if the v6 address is in the network and
(thanks emachi!) using ping6 instead of ping.

Closes-Bug: 1534578
Change-Id: Id41950f767e11884b4123fcb0bd2339636fdda68