Fix tunnel_types hieradata on compute nodes
There was a missing : in the hieradata for the compute nodes that
caused tunnel_types to not be configured.  This also made it
impossible to boot instances on tunneled networks because the port
binding always failed.

Change-Id: Icc2a45aa9514ce62497f91e6abe9261d1c1374ed
Closes-Bug: 1534349
(cherry picked from commit 11c5ab45017de16a06de7f4a7eed814dd5efcc25)