Permits configuration of Cinder enabled_backend via hieradata
It is currently possible to provide arbitrary config settings for
Cinder using the "cinder::config::cinder_config:" hiera key. To add
a backend though particular one has to edit the list of enabled
backends in Cinder too which isn't possible. This change will make
it possible using a user-customizable array of backends to be enabled.

Change-Id: Ic664c1c2b0f7b1b4b6be8b5064a38650694d4857