Remove empty value for wsrep_notify_cmd
This was being silently ignored by the mysql puppet module
prior to this commit.[1] However, now that empty values are
allowed, the overcloud deploy fails because the option
--wsrep_notify_cmd requires an argument.

This is not currently failing on master because we are
pinned to an old puppet-mysql. We will need to remove that
pin in order to get on a newer delorean repo though. Also,
this is breaking stable/liberty HA job because we use the
packaged OPM there.

[1] https://github.com/puppetlabs/puppetlabs-mysql/commit/e30e0bc958761890ea4f06cdd3f1fc7242a00fe2

Depends-On: Ifbbaf0a84dd9e0530fa8b9a42a0f66793e0df94d
Change-Id: I9e07efe1650831e81e9a783428554578874aa765
Closes-Bug: 1537720
(cherry picked from commit 403b182d94bff370d88c34d283929c133fd0a699)