Add HostnameMap to allow granular control of hostnames
Some operators desire more granular control of hostnames than is
currently possible via the *HostnameFormat parameters, in particular
mapping nodes to explicit IDs (such as inventory references) is not
easily possible.

So, add a HostnameMap parameter, which is optional and allows
explicit overriding of the default hostnames.

E.g pass an environment like this:

parameter_defaults:
  HostnameMap:
    overcloud-controller-0: overcloud-controller-prod-123-0
    overcloud-controller-1: overcloud-controller-prod-456-0
    overcloud-controller-2: overcloud-controller-prod-789-0

Note this is mapping is global (for all roles), because we
expect the keys to be unique given that they include the
role name and index by default.

Note that this depends on a fix for heat bug #1539737

Change-Id: Ib4d3d40e9523903ebccc06c3e14b2d71d924afa3
Depends-On: Ib934f443a8b8e4f75335a9d8b992e7f86791aa45
(cherry picked from commit 3c1850b780c3841ffcab1e35f81be623f0000e37)