Increase default netdev_max_backlog to 10x
It has been observed that on large clouds the netdev backlog
buffer might overflow. This change increases the default by ten
times.

The /proc/net/softnet_stat file contains a counter in the 2nd
column that is incremented when the netdev backlog queue overflows.
If this value is incrementing over time, then netdev_max_backlog
needs to be increased [1].

[1]. https://bugzilla.redhat.com/show_bug.cgi?id=1283676

Change-Id: Iec12324fd3a24e8b608b1e1849c270cc24cb0e60