Change the default value for NetworkNexusVxlanGlobalConfig
This parameter leads to the nonoperational state
in Nexus Vxlan topology when set to True, when VNIs created
but the nve peers do not get discovered on the Nexus.
It is time consuming process to debug the configuration
and find out that this parameter should be changed
to False. To prevent future problems for the future
deployment we want to default this parameter to False.

Change-Id: I685ad7d212af0d9e568acbf1ccf1607d120c195e