Add meta notify=true to rabbitmq resource
See RHBZ 1311005 and 1247303. In short: sometimes when a controller
node gets fenced, rabbitmq is unable to rejoin the cluster. To fix this
we need two steps:
1) The fix for the RA in BZ 1247303
2) Add notify=true to the meta parameters of the rabbitmq resource on
   fresh installs and updates

Note that if this change is applied on systems that do not
have the fix for the rabbitmq resource agent, no action is taken.
So when the resource agent will be updated, the notify
operation will start to work as soon as the first monitor
action will take place.

Fixes RH Bug #1311005

Change-Id: I513daf6d45e1a13d43d3c404cfd6e49d64e51d5a
(cherry picked from commit a70db5a907ad9d310a5ec890c159c92ee2122208)