Don't have separate protocols/ports for Keystone v3
The change in ab068a824ed51e78bf111387223e58e885ec5c84 is described as
temporary, so it would be better if it did not affect the EndpointMap
parameter (which is effectively a public interface, since it may be
overridden in an environment file). No configuration should end up with
different ports/protocols/hosts for Keystone v2 and v3, and somebody
customising them should not have to account for them separately. Nor
should things break when the need to distinguish between v2 and v3
endpoints goes away.

This change removes the KeystoneV3* keys from the EndpointMap input and
uses the Keystone* keys instead, so that any change to the internal
organisation becomes transparent to the user.

Closes-Bug: #1590057
Change-Id: If4cdd9232f4dbc9f2af651bbdfe68f09dc26ed2e
(cherry picked from commit d773227e7d18165750bd86a79b95d14bbc3702d9)