Move ordering constraints from httpd to openstack-core
With the move of keystone under wsgi and httpd, all openstack services received an
ordering constraint on the 'httpd' service (which now exposes keystone and horizon).

Since this is not only a layering violation, but it also removes the ability to
restart keystone (httpd) without having to restart all dependent services, we introduce
a dummy 'openstack-core' service which all other services depend on and we make also
keystone (httpd) depend on it.

The previous constraint ordering graph can be found here:
http://acksyn.org/files/tripleo/wsgi-2016-02-24-cib.pdf

Whereas after this change we have the following ordering graph:
http://acksyn.org/files/tripleo/wsgi-openstack-core.pdf

Once this is agreed upon, we can start working on fixing the upgrade path
from Liberty.

This fixes RHBZ#1290121

Closes bug: 1537885

Change-Id: Ie26908ac9bfc0b84b6b65ae3bda711236b03d9d4