Set project name for nova -> neutron notifications
This is a liberty only patch. It's a partial backport of
Ic3a329354d0ed071363183b5e06c0a42d2dd84ad, except in liberty we still
need the /v2.0 suffix on the end of auth_url in the [neutron] section in
nova.conf, so we can't backport that whole patch.

Change-Id: I317a51a12f6290a5605cf8be60d1e118cb7fb29d