Update enable-tls.yaml with new endpoints
A couple of new endpoints have been added, and if they're not in
the configured value for EndpointMap it will cause problems.

Sahara is not added as ssl-enabled because I don't believe it has
been added to the loadbalancer yet.

Note that there is work underway to CI overcloud SSL, which should
catch problems like this in the future.

Change-Id: Ia8a106fd94da7be8675ea84f5fbb9ac959771d10