Set host in nova.conf for compute nodes
In order for instance HA to function safely, nova-compute needs access
to the name by which nova knows the current compute node.

Since the names of the nova-compute and neutron ml2 agents must be the
same for vif binding to work, it also sets the host value in
neutron.conf.

Change-Id: I7d07c57b7276815c72d08acaa86f673e43eb0498