Upgrades: quiet yum upgrade on cinder nodes
Yum update on cinder nodes should be quiet, as it is on controllers,
because results of these updates are sent to Heat. I mistakenly left
this out in the first patch because i used one of the standalone node
upgrade scripts as a copy/paste base for the cinder node upgrade script.

Change-Id: Id13190dc4d242317829c7994088183f52d21461d
(cherry picked from commit d4b8297b31179f82ea3c31653802fcc1d4070e6d)