Change /etc/hosts format and add domain
Right now, the service-related IPs assosiated with the machine are
registered in the /etc/hosts with different hostnames. This is fine,
except if you need to register that hostname in a third party service
(such as FreeIPA), since the current configuration is not assigning a
domain to those IP addresses. So the current implementation requires
DNS to be properly working, which is not ideal for testing purposes.

Since the current hostnames are not currently being used; it's still
trivial to change this mapping and the format of them. instead of
having entries such as:

    <INTERNAL IP>  <node>-internalapi
    <STORAGE IP>   <node>-storage
    ...

in /etc/hosts; This changes the format to:

    <INTERNAL IP>  <node>.internalapi.<domain>  <node>.internalapi
    <STORAGE IP>   <node>.storage.<domain>  <node>.storage
    ...

So the network (external, internal, storage, etc...) is now
represented as a subdomain. For simplicity, the format without the
domain is still available through an alias.

Change-Id: I6502959a974546e5de757935acea15df6326acda