Ping retry
The single ping method in the validation script is causing for
deployments to fail. When reviewing the network connectivity, we are
finding we actually do have connectivity
( https://gist.github.com/jtaleric/0276a117625e44993be0 ). This patch is
to change the ping count from 1 to 10, to ensure the network is up.

Closes-Bug: 1563521
Change-Id: I9772407554dffa91978a49a16490ef9ed448a054