Enable galera replication for Mariadb 10.1
The generated galera config has to include additional settings for
galera to be active on MariaDB 10.1.

wsrep_on must be explicitely set to ON. On MariaDB 5.5, this was
implicitely set as soon as wsrep_provider was specified.

a valid wsrep_cluster_address must be configured in addition to
wsrep_on, otherwise recovery command mysqld_safe --wsrep-recover
cannot retrieve replication state, and cluster cannot be bootstrapped.

These explicit settings are backward compatible with MariaDB 5.5 since
the two variables exist in both versions of MariaDB.

Change-Id: I4ab4f4eeb8679899f194399ba8695155e9a2f4a5
Closes-Bug: 1563751