Add EndpointMap as output
The endpoint map contains, not only the hosts and protocols that
the resulting services will use, but also the ports. This
information is useful, and the aim of this patch is to make it
available for the tripleoclient to use it.

Change-Id: I4cc5bbf2e7200f78cd90b93659c326a9200278d7
Bug: #1569851
(cherry picked from commit 321cee7616186e83e9a5eff9e9a9abe9b9f82fab)