Add missing ManagementIpSubnet
Change 15bb67261a333f140f21208e20a112b99eeb609c added the management
network, but the ManagementIpSubnet parameter was missed being added
net-config-static-bridge.yaml. This makes the template not work by
default since Heat will error when the parameter is passed in as a
property.  This adds the parameter which fixes the template.

Change-Id: Ie745103e596c9fe2d5562acfff157157be0ff0e2