Explicitly set nova and neutron host on controllers
In I7d07c57b7276815c72d08acaa86f673e43eb0498 we set this for compute
nodes, but we also need it for controllers.  Otherwise when a
controller reboots it may come back up with a different host value,
which seems to break networking for anything that was created before
the reboot.  In my case, it changes from the short hostname (without
domain name) to the fqdn.  Since we set it to fqdn for the compute
nodes, let's do the same for controllers.

This moves all of the host setting to the base yaml of the nova
and neutron profiles.

Change-Id: Ieb793b9e9fd2dfc98584691412f9991aa99e0b47
Closes-Bug: 1604907