Retain existing ComputeHostnameFormat when upgrading older envs
As discussed at +bug/1569705 after an upgrade to liberty from an
older deployment, the hostname of the compute nodes is changed.

If we set the ComputeHostnameFormat in the upgrade-pacemaker-init
environment file parameter_defaults then we don't need to worry
about the new value being specified as default by the upgraded
overcloud.yaml since the parameter_defaults value is persisted
unless explicitly overridden with a new parameter_defaults value.

This is provided as a new environment file that needs to be
included in the first upgrade step when upgradinng from an
environment that has 'overcloud-compute-N' hostnames.

Change-Id: I2c12bd1abac65d7f13d8768f87c5ebda91164578
Related-Bug: 1569705