Fix ControllerExtraConfig parameter name
We've had a typo for a while that a parameter is named
"controllerExtraConfig" with lowercase c, which can be quite confusing
for users because the other similar parameters
(e.g. NovaComputeExtraConfig) consistently start with an upper case
letter.

We'll support both variants from now on, marking the typoed variant as
deprecated.

Change-Id: Ic67a4297e7fa08308889b95ba35389a01f70f5a4