Fix controller-no-external.yaml in bonded configs.
For some reason the controller-no-external.yaml template is configured
for DHCP on the control plane interface. We switched to static control
plane IPs before the controller-no-external.yaml was created (IIRC), so
I'm not sure how that happened. This change brings the
controller-no-external.yaml in line with the rest of the bonded NIC
templates.

Change-Id: I2ac929e241707db72a0beabf9d5cd7fc14b90f76