Aodh upgrade heat template and manifests to replace ceilometer alarms
This step needs to be run after the major-upgrade-pacemaker-init.yaml
step. It is split in three parts:

1) The first part is a shell script that disables old ceilometer alarm
pacemaker resources and removes related constraints and packages.

2) Here we run a small puppet manifest that configures aodh under httpd
via wsgi. We explicitly disable httpd's restart because by restarting
httpd manually via systemctl we risk that pacemaker's monitoring action
might happen exactly at that point and hence it will notice httpd being
down and initiate a stop/start action which will make the puppet
manifest fail. Also note that aodh uses the same backend as ceilometer.
This ensures the existing ceilometer alarms work without any need for
migration in aodh.

3) The last step is another small puppet manifest that takes
care of creating the aodh notifier, evaluator and listener pacemaker
resources and their constraints. We also restart the httpd-clone
resource to make sure that httpd is restarted with the new aodh under
wsgi configuration. This is in a separate step because we need the
guarantee that a) all packages are installed on all nodes before
creating the pacemaker resources and b) all wsgi configuration is
completed on all nodes before we restart httpd.

The rationale for this approach is that we want this
step to be verifiable on its own. That is the reason we actually
restart httpd at the very end. We want to make sure we can
verify aodh is working correctly under httpd before doing
the next upgrade steps. Note that we also include the horizon
classes in both puppet manifests in order to avoid wiping out
the horizon vhost configuration from apache.

Closes-bug: #1581504

Change-Id: I8cb4cb3193e6b823aad48ad7dbbbb227364d2a58