Don't overwrite ceph osd_journal_size
The default journal size is 5 gigs. This change stops us
overwriting it with 1 gig that is too small for production.

The config value is used by ceph only when it creates the
journal so this does not affect upgrades.

Closes-Bug: 1589909

Change-Id: I4bfd2ab47e131d8fcdd5dc75a5a56cfae8b22d5a
(cherry picked from commit 815ef9958d67e00925f13c5d23fc3898da9588e1)