Allow to manually disable post-puppet restarts
Restarting services after Puppet is vital to ensure that config changes
go applied. However, it can be sometimes desirable to prevent these
restarts to avoid downtime, if the operator is sure that no config
changes need applying. This can be a case e.g. when scaling compute
nodes. Passing the puppet-pacemaker-no-restart.yaml environment file *in
addition* to puppet-pacemaker.yaml should allow this.

This is a stop gap solution before we have proper communication between
Puppet and Pacemaker to allow selective restarts.

Change-Id: I9c3c5c10ed6ecd5489a59d7e320c3c69af9e19f4