Keystone liberty mitaka upgrade step
This step needs to be run after the major-upgrade-pacemaker-init.yaml
step and after the liberty->mitaka aodh upgrade step.
It is split in more parts:

1) Here we run a small puppet manifest that configures keystone to be
run under httpd. This part does not restart httpd at all because doing
so before turning off the openstack-keystone resource would fail
due to both services wanting to listen on the same ports.

2) In this step we add a dummy openstack-core resource, delete the
openstack-keystone resource and make sure that all the resources that
were previously depending on keystone now depend on the dummy resource.

3) We restart httpd to make sure the changes are taking place.

The rationale for doing the whole keystone migration in this single step
is that it will be verifiable on its own. We also need to add both
the aodh and the horizon apache configuration to make sure we do not
remove any services from httpd during the upgrade.

Partial-Bug: #1579705

Depends-On: I8cb4cb3193e6b823aad48ad7dbbbb227364d2a58
Depends-On: I22a348c298ff44f616b2e898f4872eddea040239

Change-Id: If85672afa8a92de2a1c0a222d28682fd100bcf7d