Drop extraconfig for neutron-nuage.yaml
This patch drops the extraconfig interface in favor
of using the composable services nested stack instead.

The benefit is that it is easier to enable multiple services
(like network and storage backends at the same time) and all
of the nuage settings get to live in the same file.

Partially-implements: blueprint composable-services-within-roles

Change-Id: I15fe14e9d6881bc408eb6bb10d9293bd914ef858