Properly direct MTU parameters and allow service defaults
This patch wires NeutronTenantMtu to path_mtu instead of the deprecated
network_device_mtu. This maintains similar behavior but allows
global_physnet_mtu to be set independently. The default for this
parameter has also been changed to an empty string, allowing proper
service defaults that enable Neutron's MTU logic to be used.

This patch also sets the default for NeutronDnsmasqOptions to an empty
string and introduces handling of empty values when creating the dnsmasq
configuration file.

Note that is a Mitaka only patch. The properties were completely removed
in Newton in change I540ba5dc69d0506f71b59746efcce94c73f9317f.

Closes-Bug: 1590100
Closes-Bug: 1590101

Change-Id: I6d12a2b741b4a81bf60f47fd553b7f71cd25b1c1