Fix network-isolation.yaml to stop clobbering Management net
The network-isolation.yaml file explicitly set the Management
network to None and the ports to noop.yaml. This resulted in the
Management network being disabled if network-management.yaml was
included before network-isolation.yaml.

These changes comment out the Management declarations to show the
default, and mention that network-management.yaml may be included.
Since the Management network is disabled by default, it isn't
necessary to noop it here. This change also places the Redis VIP
in a more logical place next to the controller ports.

Change-Id: I50a71b3ddf77bda2d213821da5e4260b503a6cb7
Fixes-bug: 1602283