L->M upgrades keystone change of path of the paste_deploy config_file
When upgrading keystone from Liberty to Mitaka the keystone-paste.ini
file moved from /usr/share/keystone to /etc/keystone.

We need to set this path during the upgrade before the convergence step
will run in order for keystone to be able to start.

While at it remove some of the older crudini calls that were only needed
during the Kilo->Liberty upgrade cycle.

Closes-Bug: #1596057

Change-Id: I5a6595583b214f93a3f70acccec483c8961a5e29