Combine BootstrapNodeDeployment with AllNodesDeployment
Currently we have a special controller-only deployment which writes
the name/ip of the "bootstrap node", e.g the cluster master, which
defaults to the first node in the Controller ResourceGroup.

Now we're moving to fully composable services/roles, it's possible
folks will want to deploy services that expect to detect the bootstrap
node (e.g so only one node does a DB sync) for non-controller roles.

So, take this opportunity to combine the bootstrap node deployment with
the "all nodes" data, such that we deploy the same data for all roles.

Because the boostrap node data is per role cluster, rather than truly
global, we pass it via input_values into each per-role Deployment.

At some future point we might consider renaming this, e.g to
something which describes per-cluster config vs "all nodes",
but as a first step let's just rationalize the resources.

Change-Id: I4011526a13c51b3d0f95c17fe8ed38115b4fdce4