Minor updates might fail with missing MysqlClustercheckPassword property
With the fixes for the galera root password:
- python-tripleoclient: https://review.openstack.org/317542
- tripleo-heat-template: https://review.openstack.org/316297

we added the MysqlClustercheckPassword property which sets the password
for the clustercheck user. When a user runs a minor update via "openstack
overcloud update" he will get the following error:

    Property error: resources[1].properties: Property MysqlClustercheckPassword not assigned

This happens because the new property does not yet have a set value.
This new property is not set unless we do an "overcloud deploy", because
the tripleoclient code generating the passwords does not get invoked
during a minor update.

We want to make sure that it has a value set. It is not used during a
minor update in any case but heat does require it.

Since the process of fixing the galera root password requires a full
overcloud deploy command anyway, it is okay to give the
MysqlClustercheckPassword property a default empty value. Upgrades are
not affected because they do run a a full overcloud deploy command.

Closes-bug: #1598890

Change-Id: I4ab82e6fba08279b96e84c48e223262f656f1120