Remove two races during the L->M Keystone under httpd migration
After adding the upgrade step to migrate keystone under httpd, we
left two small races in process:
1) The first race could result in the following error:
Graceful restart requested, doing restart (98)Address already in use: AH00072: make_sock: could not bind to address 192.0.2.14:35357

This is likely caused by removing the keystone resource and
changing constraints in a single pacemaker CIB transaction.
We are not guaranteed that pacemaker will first remove keystone
and then attempt to restart httpd due to the changed constraints.
To address this we unmanage the httpd resource before the constraint
changes and we remanage it later.

2) The second race is because after the cib-push we were not
guaranteed that the later upgrade step that reloads the httpd
configuration via 'systemctl reload httpd' was run after
httpd was started everywhere and we could get the following error:
07 07:51:14 overcloud-controller-0 systemd[1]: Unit httpd.service cannot be reloaded because it is inactive.

We add a check_resource httpd started after we remanage the httpd
resource, in order to guarantee that the httpd resource is up and
running at this point.

This should also fix RHBZ: https://bugzilla.redhat.com/show_bug.cgi?id=1348831

Closes-Bug: #1599798

Change-Id: I63871607d748068837c622fd08f2bfef120d27b7