openstack-core resource does not have interleave=true
The dummy openstack-core resource was meant to replace keystone so that
restarting keystone would not restart the whole cloud. When this
resource was introduced the paramter interleave=true was mistakenly left
out.

This causes a simple promote operation on the galera resource to restart
openstack-core and its children.

The change in master is done in puppet-tripleo because the resource
creation for the dummy resource was put there.

Change-Id: Ic590005a9419be87e6e6ea131b0ac0630c5afc19
Closes-Bug: 1603381