Fixes nova host configuration on control nodes
Closes-Bug: 1604907

nova.conf and neutron.conf is not being configured with host,
resulting in lack of fqdn in host for nova services on the control
nodes.  Upon reboot duplicate instances of nova services are found
with correct fqdn on their hostname.  This fixes nova and neutron
conf to set the host value in controller manifests, similar to how
it is done in compute manifest.

This is a backport of Ieb793b9e9fd2dfc98584691412f9991aa99e0b47

Change-Id: Ie11b39de35194c020a8a09b57fdb11e882c5ec7b
Signed-off-by: Tim Rozet <trozet@redhat.com>