Include new apache pacemaker profile
In https://review.openstack.org/345013 we are working to create a new
profile for Apache with Pacemaker. We can't include this class in the
profiles (keystone, etc) yet because it would cause a duplicate resource
error in Puppet catalog.

So we need to first clean the code in overcloud_controller_pacemaker.pp
and include the new class, and then we can either create a new
composable service in THT for apache or include the class directly in
puppet-tripleo profiles that need it (keystone, horizon, etc).

Change-Id: I1222fbf099b4268e33ec8ee32380f0a6b1d5e1af
Depends-On: I9ae6cee2bfb0f8974d41d700454cfde2df06c2d1